+++ src/main_area_dna.rs
@@ -66,6 +66,7 @@ mod feature_tree_ui;
 #[path = "main_area_dna/formula_controls.rs"]
 mod formula_controls;
 
+mod conservation_request_ui;
 #[path = "main_area_dna/genomic_regions_ui.rs"]
 mod genomic_regions_ui;
 mod locus_inspector;
@@ -1730,7 +1731,7 @@ pub struct MainAreaDna {
     genomic_region_conservation_set_id: String,
     genomic_region_conservation_region_id: String,
     genomic_region_conservation_expected_sha256: Option<String>,
-    genomic_region_conservation_query_genome_id: String,
+    genomic_region_conservation_request: gentle_protocol::GenomicRegionHomologyScreenRequest,
     genomic_region_conservation_report:
         Option<Arc<gentle_protocol::GenomicRegionHomologyScreenReport>>,
     genomic_region_conservation_selected_block_id: Option<String>,
@@ -2607,7 +2608,7 @@ impl MainAreaDna {
             genomic_region_conservation_set_id: String::new(),
             genomic_region_conservation_region_id: String::new(),
             genomic_region_conservation_expected_sha256: None,
-            genomic_region_conservation_query_genome_id: String::new(),
+            genomic_region_conservation_request: Default::default(),
             genomic_region_conservation_report: None,
             genomic_region_conservation_selected_block_id: None,
             genomic_region_conservation_evidence_region_ids: BTreeSet::new(),